OPENSTACK-2533: remove agent manager
